python-backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.10.6

# Use compatible versions - langchain-community 0.3.27 requires langchain-core>=0.3.66
langchain-core==0.3.76
langchain-anthropic==0.3.5
langchain-community==0.3.27

# RAG dependencies
chromadb==0.5.23
sentence-transformers==3.0.1

# keep httpx below the breaking change that caused "unexpected keyword argument 'proxies'"
httpx==0.27.2
